requirements.txt
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.7.0
eventlet==0.38.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.1
requests==2.31.0
setuptools==75.6.0
simple-websocket==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==7.2
